refactor: IPNE 描画の敵層を drawEnemies へ分離
- 敵スプライト・撃破アニメ・ボスHPオーラ・HPバー・攻撃エフェクト描画を逐語移植
- getEnemyStateFrame ヘルパーを drawEnemies.ts へ移動
- 特性化テストで描画コマンド列の不変を確認（2 passed）
