+++ src/features/ipne/presentation/screens/render/renderGameFrame.ts
@@ -9,7 +9,6 @@ import {
   calculateViewport,
   calculateTileSize,
   getCanvasSize,
-  EnemyState,
   EnemyType,
   drawAutoMap,
   findPath,
@@ -20,40 +19,23 @@ import { SPRITE_SIZES } from '../../config';
 import {
   EffectType,
   calculatePowerLevel,
-  getDeathPhase,
-  getDeathScale,
-  isDeathAnimationComplete,
   shouldTriggerWarning,
   getWarningPhase,
-  getBossAuraConfig,
   BOSS_WARNING_DURATION,
 } from '../../effects';
 import { isComboActive, COMBO_DISPLAY_MIN } from '../../../domain/services/comboService';
 import {
-  SpriteDefinition,
   FLOOR_SPRITE,
   WALL_SPRITE,
   getStageFloorSprite,
   getStageWallSprite,
   getPlayerSpriteSheet,
-  getEnemySpriteSheet,
-  ATTACK_SLASH_SPRITE_SHEET,
   WARRIOR_ATTACK_SPRITE_SHEETS,
   THIEF_ATTACK_SPRITE_SHEETS,
   WARRIOR_DAMAGE_SPRITES,
   THIEF_DAMAGE_SPRITES,
   WARRIOR_IDLE_SPRITE_SHEETS,
   THIEF_IDLE_SPRITE_SHEETS,
-  PATROL_ATTACK_FRAME,
-  CHARGE_RUSH_FRAME,
-  RANGED_CAST_FRAME,
-  SPECIMEN_MUTATE_FRAME,
-  BOSS_ATTACK_FRAME,
-  BOSS_DAMAGE_FRAME,
-  MINI_BOSS_ATTACK_FRAME,
-  MINI_BOSS_DAMAGE_FRAME,
-  MEGA_BOSS_ATTACK_FRAME,
-  MEGA_BOSS_DAMAGE_FRAME,
 } from '../../sprites';
 import { drawPlayerAura } from '../../effects/aura';
 import { drawWeaponTrail, getWeaponTier, WeaponTier, drawShockwave } from '../../effects/weaponEffect';
@@ -62,29 +44,7 @@ import { getStageIntroPhase, getStageIntroAlpha, getStageIntroTextAlpha, getGame
 import type { RenderContext, FrameContext } from './renderContext';
 import type { Position, Viewport } from '../../../index';
 import { drawWorld } from './drawWorld';
-
-/** 敵の状態に応じた特殊フレームを返す（Phase 3） */
-function getEnemyStateFrame(enemyType: string, enemyState: string): SpriteDefinition | null {
-  if (enemyState === EnemyState.ATTACK) {
-    switch (enemyType) {
-      case EnemyType.PATROL: return PATROL_ATTACK_FRAME;
-      case EnemyType.CHARGE: return CHARGE_RUSH_FRAME;
-      case EnemyType.RANGED: return RANGED_CAST_FRAME;
-      case EnemyType.SPECIMEN: return SPECIMEN_MUTATE_FRAME;
-      case EnemyType.BOSS: return BOSS_ATTACK_FRAME;
-      case EnemyType.MINI_BOSS: return MINI_BOSS_ATTACK_FRAME;
-      case EnemyType.MEGA_BOSS: return MEGA_BOSS_ATTACK_FRAME;
-    }
-  }
-  if (enemyState === EnemyState.KNOCKBACK) {
-    switch (enemyType) {
-      case EnemyType.BOSS: return BOSS_DAMAGE_FRAME;
-      case EnemyType.MINI_BOSS: return MINI_BOSS_DAMAGE_FRAME;
-      case EnemyType.MEGA_BOSS: return MEGA_BOSS_DAMAGE_FRAME;
-    }
-  }
-  return null;
-}
+import { drawEnemies } from './drawEnemies';
 
 /**
  * ゲームフレームを描画する
@@ -224,111 +184,8 @@ export function renderGameFrame(rc: RenderContext): void {
   // ワールド描画（背景・マップ・パス・罠・壁・アイテム）
   drawWorld(frame, shakeOffset);
 
-  // 敵描画（T-02.3: スプライト描画）
-  for (const enemy of enemies) {
-    if (
-      enemy.x < viewport.x - 1 ||
-      enemy.x > viewport.x + viewport.width + 1 ||
-      enemy.y < viewport.y - 1 ||
-      enemy.y > viewport.y + viewport.height + 1
-    ) {
-      if (!useFullMap) continue;
-    }
-
-    const enemyScreen = toScreenPosition(enemy);
-    const enemySpriteSize =
-      enemy.type === EnemyType.MEGA_BOSS ? SPRITE_SIZES.megaBoss :
-      enemy.type === EnemyType.BOSS ? SPRITE_SIZES.boss :
-      enemy.type === EnemyType.MINI_BOSS ? SPRITE_SIZES.miniBoss :
-      SPRITE_SIZES.base;
-    const enemyDrawSize = enemySpriteSize * spriteScale;
-    const enemyDrawX = enemyScreen.x - enemyDrawSize / 2;
-    const enemyDrawY = enemyScreen.y - enemyDrawSize / 2;
-
-    // 撃破アニメーション中の描画
-    if (enemy.isDying && enemy.deathStartTime) {
-      const elapsed = now - enemy.deathStartTime;
-      if (isDeathAnimationComplete(elapsed)) continue;
-
-      const phase = getDeathPhase(elapsed);
-      const scale = getDeathScale(elapsed);
-
-      if (phase === 1) {
-        // フェーズ1: 縮小描画（100ms）
-        ctx.save();
-        ctx.translate(enemyScreen.x, enemyScreen.y);
-        ctx.scale(scale, scale);
-        const scaledDrawX = -enemyDrawSize / 2;
-        const scaledDrawY = -enemyDrawSize / 2;
-        const enemySheet = getEnemySpriteSheet(enemy.type);
-        spriteRenderer.drawSprite(ctx, enemySheet.sprites[0], scaledDrawX, scaledDrawY, spriteScale);
-        ctx.restore();
-      } else if (phase === 2) {
-        // フェーズ2: 白フラッシュ（50ms）
-        ctx.save();
-        ctx.globalAlpha = 0.8;
-        ctx.fillStyle = '#ffffff';
-        ctx.beginPath();
-        ctx.arc(enemyScreen.x, enemyScreen.y, enemyDrawSize / 2, 0, Math.PI * 2);
-        ctx.fill();
-        ctx.restore();
-      }
-      // フェーズ3: スプライト非表示（パーティクルのみ）
-      continue;
-    }
-
-    // ボスHP残量オーラ描画
-    const isBossType = enemy.type === EnemyType.BOSS ||
-      enemy.type === EnemyType.MINI_BOSS || enemy.type === EnemyType.MEGA_BOSS;
-    if (isBossType && enemy.hp > 0) {
-      const hpRatio = enemy.maxHp > 0 ? enemy.hp / enemy.maxHp : 1;
-      const auraConfig = getBossAuraConfig(hpRatio);
-      if (auraConfig) {
-        const pulseT = (now % auraConfig.pulsePeriod) / auraConfig.pulsePeriod;
-        const pulseAlpha = 0.15 + 0.15 * Math.sin(pulseT * Math.PI * 2);
-        ctx.save();
-        ctx.globalAlpha = pulseAlpha;
-        const gradient = ctx.createRadialGradient(
-          enemyScreen.x, enemyScreen.y, enemyDrawSize * 0.3,
-          enemyScreen.x, enemyScreen.y, enemyDrawSize * 0.8
-        );
-        gradient.addColorStop(0, 'rgba(220, 38, 38, 0.6)');
-        gradient.addColorStop(1, 'rgba(220, 38, 38, 0)');
-        ctx.fillStyle = gradient;
-        ctx.fillRect(
-          enemyScreen.x - enemyDrawSize,
-          enemyScreen.y - enemyDrawSize,
-          enemyDrawSize * 2,
-          enemyDrawSize * 2
-        );
-        ctx.restore();
-      }
-    }
-
-    const blinkOff = enemy.state === EnemyState.KNOCKBACK && Math.floor(now / 100) % 2 === 1;
-    if (blinkOff) continue;
-
-    const enemySheet = getEnemySpriteSheet(enemy.type);
-
-    // 敵状態別フレーム選択（Phase 3）
-    const enemyStateFrame = getEnemyStateFrame(enemy.type, enemy.state);
-    if (enemyStateFrame) {
-      spriteRenderer.drawSprite(ctx, enemyStateFrame, enemyDrawX, enemyDrawY, spriteScale);
-    } else {
-      spriteRenderer.drawAnimatedSprite(ctx, enemySheet, now, enemyDrawX, enemyDrawY, spriteScale);
-    }
-  }
-
-  // 攻撃エフェクト描画（T-02.8: 斬撃アニメーション）
-  if (attackEffect && now < attackEffect.until) {
-    const effectPos = attackEffect.position;
-    const screen = toScreenPosition(effectPos);
-    const slashDrawSize = SPRITE_SIZES.base * spriteScale;
-    const slashDrawX = screen.x - slashDrawSize / 2;
-    const slashDrawY = screen.y - slashDrawSize / 2;
-
-    spriteRenderer.drawAnimatedSprite(ctx, ATTACK_SLASH_SPRITE_SHEET, now, slashDrawX, slashDrawY, spriteScale);
-  }
+  // 敵描画（敵スプライト・撃破アニメ・ボスHPオーラ・攻撃エフェクト）
+  drawEnemies(frame);
 
   // パーティクルエフェクトシステム
   const em = effectManagerRef.current;
